Updated min and max

+++ src/numpp.hpp
@@ -2,7 +2,7 @@
 #define NUMPP_H
 
 #include <vector>
-
+#include<string>
 using namespace std;
 
 template <class T>
@@ -11,15 +11,15 @@ class numpp {
     vector<int> dim;
     vector<T> data;
     bool is_zero;
-    uint64_t ndim;
+    long long int ndim;
     int shape_match(vector<int>);
   public:
     numpp(vector<int>);
     void shape();
     void reshape(vector<int>);
     void push_vec(vector<int>, vector<T>);
-    uint64_t max();
-    uint64_t min();
+    long long int max();
+    long long int min();
     void zeros();
     void print();
     void add(numpp,numpp);
